+++ backend/src/main/java/com/github/mysterix5/vover/VoverApplication.java
@@ -1,13 +1,13 @@
-package com.github.mysterix5.capstone;
+package com.github.mysterix5.vover;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class CapstoneApplication {
+public class VoverApplication {
 
 	public static void main(String[] args) {
-		SpringApplication.run(CapstoneApplication.class, args);
+		SpringApplication.run(VoverApplication.class, args);
 	}
 
 }
