+++ backend/src/main/java/com/github/mysterix5/vover/model/Availability.java
@@ -1,4 +1,4 @@
-package com.github.mysterix5.capstone.model;
+package com.github.mysterix5.vover.model;
 
 public enum Availability {
     ABSENT, PUBLIC, INVALID
