+++ backend/src/main/java/com/github/mysterix5/vover/model/WordResponseDTO.java
@@ -1,4 +1,4 @@
-package com.github.mysterix5.capstone.model;
+package com.github.mysterix5.vover.model;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
