+++ backend/src/main/java/com/github/mysterix5/vover/textHandling/WordsMongoRepository.java
@@ -1,6 +1,6 @@
-package com.github.mysterix5.capstone.textHandling;
+package com.github.mysterix5.vover.textHandling;
 
-import com.github.mysterix5.capstone.model.WordDbEntity;
+import com.github.mysterix5.vover.model.WordDbEntity;
 import org.springframework.data.mongodb.repository.MongoRepository;
 
 public interface WordsMongoRepository extends MongoRepository<WordDbEntity, String> {
