+++ backend/src/test/java/com/github/mysterix5/vover/addWords/AddWordServiceTest.java
@@ -1,9 +1,9 @@
-package com.github.mysterix5.capstone.addWords;
+package com.github.mysterix5.vover.addWords;
 
-import com.github.mysterix5.capstone.model.WordDbEntity;
-import com.github.mysterix5.capstone.model.WordInput;
-import com.github.mysterix5.capstone.model.WordTag;
-import com.github.mysterix5.capstone.textHandling.WordsMongoRepository;
+import com.github.mysterix5.vover.model.WordDbEntity;
+import com.github.mysterix5.vover.model.WordInput;
+import com.github.mysterix5.vover.model.WordTag;
+import com.github.mysterix5.vover.textHandling.WordsMongoRepository;
 import org.junit.jupiter.api.Test;
 import org.mockito.Mockito;
